UNR-2129 SpatialWorkerCleanup (#1610)
* Moved the StaticComponetView and the GlobalStateManager to the SpatialGameInstance

* Update SpatialGDK/Source/SpatialGDK/Private/EngineClasses/SpatialNetDriver.cpp


* Added handling of connection error

* Addressing PR feedback

* Fixed the build

* Addressing PR feedback

* GSM and StaticComponentView are now created as part of USpatialGameInstance::Init

* Added nullptr checks

* Update SpatialGDK/Source/SpatialGDK/Private/EngineClasses/SpatialNetDriver.cpp


* Moved GSM and StaticComponentView to private in GameInstance, added getters
