+++ SpatialGDK/Source/SpatialGDK/Public/EngineClasses/SpatialGameInstance.h
@@ -9,6 +9,8 @@
 
 class USpatialLatencyTracer;
 class USpatialWorkerConnection;
+class UGlobalStateManager;
+class USpatialStaticComponentView;
 
 DECLARE_LOG_CATEGORY_EXTERN(LogSpatialGameInstance, Log, All);
 
@@ -42,6 +44,8 @@ public:
 
 	FORCEINLINE USpatialWorkerConnection* GetSpatialWorkerConnection() { return SpatialConnection; }
 	FORCEINLINE USpatialLatencyTracer* GetSpatialLatencyTracer() { return SpatialLatencyTracer; }
+	FORCEINLINE UGlobalStateManager* GetGlobalStateManager() { return GlobalStateManager; };
+	FORCEINLINE USpatialStaticComponentView* GetStaticComponentView() { return StaticComponentView; };
 
 	void HandleOnConnected();
 	void HandleOnConnectionFailed(const FString& Reason);
@@ -72,4 +76,13 @@ private:
 
 	UPROPERTY()
 	USpatialLatencyTracer* SpatialLatencyTracer = nullptr;
+
+	// GlobalStateManager must persist when server traveling
+	UPROPERTY()
+	UGlobalStateManager* GlobalStateManager;
+
+	// StaticComponentView must persist when server traveling
+	UPROPERTY()
+	USpatialStaticComponentView* StaticComponentView;
+
 };
