+++ SpatialGDK/Source/SpatialGDK/Public/Interop/Connection/SpatialWorkerConnection.h
@@ -20,11 +20,6 @@
 
 DECLARE_LOG_CATEGORY_EXTERN(LogSpatialWorkerConnection, Log, All);
 
-class UGlobalStateManager;
-class USpatialGameInstance;
-class USpatialStaticComponentView;
-class UWorld;
-
 enum class ESpatialConnectionType
 {
 	Receptionist,
@@ -38,8 +33,6 @@ class SPATIALGDK_API USpatialWorkerConnection : public UObject, public FRunnable
 	GENERATED_BODY()
 
 public:
-	void Init(USpatialGameInstance* InGameInstance);
-
 	virtual void FinishDestroy() override;
 	void DestroyConnection();
 
@@ -83,19 +76,12 @@ public:
 	DECLARE_DELEGATE_TwoParams(OnConnectionToSpatialOSFailedDelegate, uint8_t, const FString&);
 	OnConnectionToSpatialOSFailedDelegate OnFailedToConnectCallback;
 
-	UPROPERTY()
-	USpatialStaticComponentView* StaticComponentView;
-
-	UPROPERTY()
-	UGlobalStateManager* GlobalStateManager;
-
 private:
 	void ConnectToReceptionist(uint32 PlayInEditorID);
 	void ConnectToLocator();
 	void FinishConnecting(Worker_ConnectionFuture* ConnectionFuture);
 
 	void OnConnectionSuccess();
-	void OnPreConnectionFailure(const FString& Reason);
 	void OnConnectionFailure();
 
 	ESpatialConnectionType GetConnectionType() const;
@@ -122,8 +108,6 @@ private:
 private:
 	Worker_Connection* WorkerConnection;
 	Worker_Locator* WorkerLocator;
-
-	TWeakObjectPtr<USpatialGameInstance> GameInstance;
 
 	bool bIsConnected;
 	bool bConnectAsClient = false;
